refactor(moderation): split live types out of dead ContentModerationService (#5047)
The ContentModerationService class is never constructed — only the
top-level value types in its file are live (ContentFilterReason,
ContentSeverity, ModerationResult, MuteListEntry, used by the report
and content-warning flows). Move those four to content_moderation_types.dart,
delete the 700-line dead service, and repoint the 13 import sites.
NostrListServiceMixin stays (bookmark_service still uses it).
