+++ mobile/lib/blocs/dm/conversation_actions/conversation_actions_cubit.dart
@@ -6,7 +6,7 @@ import 'package:content_blocklist_repository/content_blocklist_repository.dart';
 import 'package:dm_repository/dm_repository.dart';
 import 'package:equatable/equatable.dart';
 import 'package:openvine/observability/reportable_error.dart';
-import 'package:openvine/services/content_moderation_service.dart';
+import 'package:openvine/services/content_moderation_types.dart';
 import 'package:openvine/services/content_reporting_service.dart';
 
 enum ConversationActionsStatus { idle, processing, success, failure }
