+++ mobile/lib/l10n/content_filter_reason_localizations.dart
@@ -1,5 +1,5 @@
 import 'package:openvine/l10n/generated/app_localizations.dart';
-import 'package:openvine/services/content_moderation_service.dart';
+import 'package:openvine/services/content_moderation_types.dart';
 
 extension ContentFilterReasonLocalizations on AppLocalizations {
   String reportReasonTitle(ContentFilterReason reason) {
